:truck: chore: refactored mappers and core data
