+++ app/src/main/java/com/shub39/grit/core/data/BootReceiver.kt
@@ -20,6 +20,7 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.util.Log
+import com.shub39.grit.core.data.notification.NotificationAlarmScheduler
 import com.shub39.grit.core.habits.domain.HabitRepo
 import com.shub39.grit.core.tasks.domain.TaskRepo
 import kotlinx.coroutines.CoroutineScope
