+++ app/src/main/java/com/shub39/grit/habits/data/Mappers.kt
@@ -14,16 +14,12 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.grit.core.data
+package com.shub39.grit.habits.data
 
 import com.shub39.grit.core.habits.domain.Habit
 import com.shub39.grit.core.habits.domain.HabitStatus
-import com.shub39.grit.core.tasks.domain.Category
-import com.shub39.grit.core.tasks.domain.Task
 import com.shub39.grit.habits.data.database.HabitEntity
 import com.shub39.grit.habits.data.database.HabitStatusEntity
-import com.shub39.grit.tasks.data.database.CategoryEntity
-import com.shub39.grit.tasks.data.database.TaskEntity
 
 fun HabitEntity.toHabit(): Habit {
     return Habit(
@@ -55,34 +51,4 @@ fun Habit.toHabitEntity(): HabitEntity {
 
 fun HabitStatus.toHabitStatusEntity(): HabitStatusEntity {
     return HabitStatusEntity(id = id, habitId = habitId, date = date)
-}
-
-fun Task.toTaskEntity(): TaskEntity {
-    return TaskEntity(
-        id = id,
-        categoryId = categoryId,
-        title = title,
-        index = index,
-        status = status,
-        reminder = reminder,
-    )
-}
-
-fun TaskEntity.toTask(): Task {
-    return Task(
-        id = id,
-        categoryId = categoryId,
-        title = title,
-        index = index,
-        status = status,
-        reminder = reminder,
-    )
-}
-
-fun CategoryEntity.toCategory(): Category {
-    return Category(id = id, name = name, index = index, color = color)
-}
-
-fun Category.toCategoryEntity(): CategoryEntity {
-    return CategoryEntity(id = id, name = name, color = color, index = index)
 }
