+++ app/src/main/java/com/shub39/grit/habits/data/repository/HabitRepository.kt
@@ -16,11 +16,7 @@
  */
 package com.shub39.grit.habits.data.repository
 
-import com.shub39.grit.core.data.GritNotificationManager
-import com.shub39.grit.core.data.toHabit
-import com.shub39.grit.core.data.toHabitEntity
-import com.shub39.grit.core.data.toHabitStatus
-import com.shub39.grit.core.data.toHabitStatusEntity
+import com.shub39.grit.core.data.notification.GritNotificationManager
 import com.shub39.grit.core.domain.SettingsDatastore
 import com.shub39.grit.core.habits.domain.Habit
 import com.shub39.grit.core.habits.domain.HabitRepo
@@ -30,6 +26,10 @@ import com.shub39.grit.core.habits.domain.OverallAnalytics
 import com.shub39.grit.core.now
 import com.shub39.grit.habits.data.database.HabitStatusDao
 import com.shub39.grit.habits.data.database.HabitsDao
+import com.shub39.grit.habits.data.toHabit
+import com.shub39.grit.habits.data.toHabitEntity
+import com.shub39.grit.habits.data.toHabitStatus
+import com.shub39.grit.habits.data.toHabitStatusEntity
 import kotlin.time.ExperimentalTime
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
