+++ app/src/main/java/com/shub39/grit/tasks/data/Mappers.kt
@@ -1,0 +1,52 @@
+/*
+ * Copyright (C) 2026  Shubham Gorai
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package com.shub39.grit.tasks.data
+
+import com.shub39.grit.core.tasks.domain.Category
+import com.shub39.grit.core.tasks.domain.Task
+import com.shub39.grit.tasks.data.database.CategoryEntity
+import com.shub39.grit.tasks.data.database.TaskEntity
+
+fun Task.toTaskEntity(): TaskEntity {
+    return TaskEntity(
+        id = id,
+        categoryId = categoryId,
+        title = title,
+        index = index,
+        status = status,
+        reminder = reminder,
+    )
+}
+
+fun TaskEntity.toTask(): Task {
+    return Task(
+        id = id,
+        categoryId = categoryId,
+        title = title,
+        index = index,
+        status = status,
+        reminder = reminder,
+    )
+}
+
+fun CategoryEntity.toCategory(): Category {
+    return Category(id = id, name = name, index = index, color = color)
+}
+
+fun Category.toCategoryEntity(): CategoryEntity {
+    return CategoryEntity(id = id, name = name, color = color, index = index)
+}
